requirements.txt
flask
opencv-python
numpy
Pillow
matplotlib
